Update README.md
## Description
Running $`yo studio-app` searches for an existing `studio-app` plugin and returns the err:
```
You don't seem to have a generator with the name “studio-app” installed.
```
Simply running $`yo` instead provides a nice CLI experience and allows the user to just select the available `Studio Plugin` option under the `Run a generator` list. This PR updates the README and directs the user to do that.